Issue #11163: Enforce filesize LinelengthSimple
